support streaming for nl search (#164)
