+++ client/src/pages/Search/index.tsx
@@ -51,7 +51,6 @@ let onboardingFinished = false;
 const SearchPage = () => {
   const { setInputValue, globalRegex, searchType, setSearchType } =
     useContext(SearchContext);
-  const { isRepoManagementAllowed } = useContext(DeviceContext);
   const [shouldShowWelcome, setShouldShowWelcome] = useState(
     !getPlainFromStorage(ONBOARDING_DONE_KEY),
   );
@@ -60,6 +59,7 @@ const SearchPage = () => {
     searchQuery: nlSearchQuery,
     data: nlData,
     loading: nlLoading,
+    query: nlQuery,
   } = useSearch<NLSearchResponse>();
   const { updateCurrentTabName } = useContext(TabsContext);
 
@@ -201,12 +201,13 @@ const SearchPage = () => {
             loading={nlLoading}
             resultsData={nlData}
             handleRetry={handleRetry}
+            nlQuery={nlQuery}
           />
         );
       default:
         return <HomePage />;
     }
-  }, [data, loading, nlLoading, nlData, handleRetry, navigatedItem]);
+  }, [data, loading, nlLoading, nlData, handleRetry, navigatedItem, nlQuery]);
 
   return shouldShowWelcome ? (
     <div className="text-gray-200">
